Make quickjs.h -Wall -Wextra -pedantic clean
Fixes: https://github.com/quickjs-ng/quickjs/issues/585
